# Hash estimator/transformer applicable to all numerics and bool