… test/Microsoft.ML.Tests/Transformers/HashTests.cs