Fixes #1031. Hashes all numeric types, and boolean types. "Time" types are omitted, because I don't anticipate great desire for them, but they could potentially be more easily added.

I engaged in some trickery around the JIT's handlings of generics and value types in order to maintain efficiency of the previous "explicit" implementation. Please see the comment starting line 526 in the last commit (as of the time of writing) for more info.

There are to start three commits, that are deliberately structured in a way so as to be useful:

1. Introduction of benchmarks,
2. Simplification of code,
3. Expanding of types.

The benchmarking is somewhat interesting. The variation between different runs of the benchmarks was all over the place with the old code (for some reason), but not on the new code for some reason. I'm not quite sure why. Anyway these were the first runs from each.

## Before

|            Method |        Mean |      Error |     StdDev |
| ----------------- |------------ |----------- |----------- |
|  HashScalarString |  4,738.1 us | 146.250 us | 168.421 us |
|   HashScalarFloat |  1,172.0 us |  30.922 us |  34.370 us |
|  HashScalarDouble |  1,535.2 us |  28.194 us |  26.373 us |
|     HashScalarKey |    965.2 us |  18.733 us |  18.398 us |
|  HashVectorString | 28,295.6 us | 554.444 us | 569.373 us |
|   HashVectorFloat | 10,442.6 us |   6.021 us |   5.338 us |
|  HashVectorDouble | 12,501.8 us | 240.539 us | 267.359 us |
|     HashVectorKey |  9,839.5 us | 270.684 us | 265.848 us |

## After

|            Method |        Mean |      Error |     StdDev |
| -----------------:|------------:|-----------:|-----------:|
|  HashScalarString |  4,769.2 us | 125.379 us | 144.387 us |
|   HashScalarFloat |  1,094.3 us |   4.808 us |   3.754 us |
|  HashScalarDouble |  1,529.9 us |  33.315 us |  29.533 us |
|     HashScalarKey |    960.3 us |  17.989 us |  19.248 us |
|  HashVectorString | 28,513.9 us | 520.837 us | 487.191 us |
|   HashVectorFloat |  9,932.0 us | 231.398 us | 257.198 us |
|  HashVectorDouble | 12,264.1 us | 286.483 us | 329.915 us |
|     HashVectorKey |  9,118.8 us | 178.890 us | 167.334 us ||