@Zruty0 2018-10-19
If you don't feel like doing it in this PR, then please keep #1031 open until it's done.

---
In reply to: [431420698](https://github.com/dotnet/machinelearning/pull/1303#issuecomment-431420698) [](ancestors = 431420698)