@TomFinley 2018-10-19
> If you don't feel like doing it in this PR, then please keep #1031 open until it's done.

Actually I've tried to do this, and it's surprisingly thorny, and in ways that make me think that it actually deserves its own issue -- sparsity, handling of defaults, and whatnot are the major things that surprised me with this one, and I don't know that #1031 is the best place.