@sfilipi 2018-10-19
not the best person to review, perhaps, but nothing looks wrong to me .

---
In reply to: [431498395](https://github.com/dotnet/machinelearning/pull/1303#issuecomment-431498395) [](ancestors = 431498395)